+++ app/controllers/api/v1/coordinates_controller.rb
@@ -1,12 +1,20 @@
 class Api::V1::CoordinatesController < ApplicationController
-  before_action :set_coordinate, only: %i[show destroy]
+  include ResourceFinder
+  before_action :set_coordinate, only: %i[show update destroy]
+  skip_before_action :authenticate_request, only: %i[show]
+  before_action :authorize_policy
 
   def show
+    authorize @coordinate
+
     render json: @coordinate
   end
 
   def create
-    @coordinate = Coordinate.new(coordinate_params)
+    @coordinate = parentable.coordinates.new(coordinate_params)
+
+    authorize @coordinate
+
     if @coordinate.save
       render json: @coordinate, status: :created
     else
@@ -15,6 +23,8 @@ class Api::V1::CoordinatesController < ApplicationController
   end
 
   def update
+    authorize @coordinate
+
     if @coordinate.update(coordinate_params)
       render json: { status: 'Update', data: @coordinate }, status: :ok
     else
@@ -23,6 +33,8 @@ class Api::V1::CoordinatesController < ApplicationController
   end
 
   def destroy
+    authorize @coordinate
+    
     if @coordinate.destroy!
       render json: { status: 'Delete' }, status: :no_content
     else
@@ -33,13 +45,14 @@ class Api::V1::CoordinatesController < ApplicationController
   private
 
   def set_coordinate
-    @coordinate = Coordinate.find(params[:id])
-  rescue ActiveRecord::RecordNotFound => e
-    logger.info e
-    render json: { message: 'coordinate id not found' }, status: :not_found
+    @coordinate = parentable.coordinates
   end
 
   def coordinate_params
-    params.require(:coordinate).permit(:coordinatable_type, :coordinatable_id)
+    params.permit(:latitude, :longitude)
+  end
+
+  def authorize_policy
+    authorize Coordinate
   end
 end
